+++ changelog.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Changes
-- fixed recipe to upgrade sack into lunchbox, now it properly makes items stay
+- fixed recipe to upgrade sack into lunchbox, now it properly makes items stay and refunds incompatible items.
+- fixed recipe to upgrade sack into toolbelt, now it properly makes items stay and refunds incompatible items.
 
 ## [0.0.33]
 ### Changes